Move `graphgen` to new module: `graphfrag.py`
